requirements.txt
requests
beautifulsoup4
python-dotenv
selenium==4.19.0
webdriver-manager==4.0.1
python-dotenv==1.0.1
